support symbolic links
